+++ src/component/atom/common/CheckSign/CheckSign.stories.tsx
