+++ src/component/atom/common/DeleteButton/DeleteButton.stories.tsx
