+++ src/component/atom/common/DeleteButton/DeleteButton.tsx
