+++ src/component/molecule/common/SpellingChecker/SpellingChecker.stories.tsx
